requirements.txt
torch
facenet-pytorch
Pillow
numpy
tqdm
